Cleaning up the ME GUIs (Items+Fluids) (#5105)
Completely refactor all UI. Introduce JSON based definitions for screens (can be overridden in resource packs). Unifies fluid and item terminal UI code. Serializes item data from server->client using "share tags" (i.e. don't include the content of bags if they support that), and no longer synchronize capability data to the client. Use "serial numbers" to refer back to items on the server when a player interacts with items in terminals. This should fix #5083 "for good" while still using less network bandwidth. It should also fix #5199 because serialization of items into packets is rewritten and now uses a length-prefix, reducing the chance for corrupted packets due to very long item NBT data.
Rewrote crafting status / network status to use custom packets rather than reusing the inventory sync packets.

Also implements #3787, #3922

Also adds:
    Start using share tags / stop sending caps to clients, which should reduce bandwidth usage when opening terminals
    Consider missing items on "shift-clickicking" recipes into the crafting terminal from JEI
    Adds tooltip indicating which upgrades are compatible to the upgrade panel for various machines
    Adds information on how priority works to the priority screen
    Use standard slots for the player inventory, which should fix most inventory sorting mods moving the hotbar into the main inventory when sorting the player inventory within an AE2 screen.
    Add clickable "show recipes" areas that open the appropriate JEI category to the inscriber and grindstone

+++ src/main/java/appeng/container/guisync/GuiSync.java
@@ -30,5 +30,5 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 public @interface GuiSync {
 
-    int value();
+    short value();
 }
